+++ packages/graphql-cli-codegen-plugin/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @graphql-cli/codegen
 
+## 1.17.16
+
+### Patch Changes
+
+- Updated dependencies [920d8e95]
+  - @graphql-codegen/cli@1.19.4
+
+## 1.17.15
+
+### Patch Changes
+
+- 1183d173: Bump all packages to resolve issues with shared dependencies
+- Updated dependencies [1183d173]
+  - @graphql-codegen/cli@1.19.3
+
+## 1.17.14
+
+### Patch Changes
+
+- Updated dependencies [faa13973]
+- Updated dependencies [faa13973]
+  - @graphql-codegen/cli@1.19.2
+
 ## 1.17.13
 
 ### Patch Changes