+++ packages/plugins/typescript/typescript/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @graphql-codegen/typescript
 
+## 1.19.0
+
+### Minor Changes
+
+- 1d6a593f: Added `useImplementingTypes` flag for generating code that uses implementing types instead of interfaces
+
+### Patch Changes
+
+- Updated dependencies [8356f8a2]
+  - @graphql-codegen/visitor-plugin-common@1.17.21
+
+## 1.18.1
+
+### Patch Changes
+
+- 1183d173: Bump all packages to resolve issues with shared dependencies
+- Updated dependencies [1183d173]
+  - @graphql-codegen/visitor-plugin-common@1.17.20
+  - @graphql-codegen/plugin-helpers@1.18.2
+
+## 1.18.0
+
+### Minor Changes
+
+- 49242c20: Added a "defaultValue" option in the "avoidOptionals" config
+  See https://github.com/dotansimha/graphql-code-generator/issues/5112
+
+### Patch Changes
+
+- Updated dependencies [99819bf1]
+- Updated dependencies [c3b59e81]
+  - @graphql-codegen/visitor-plugin-common@1.17.19
+
 ## 1.17.11
 
 ### Patch Changes