+++ packages/utils/plugins-helpers/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @graphql-codegen/plugin-helpers
 
+## 1.18.2
+
+### Patch Changes
+
+- 1183d173: Bump all packages to resolve issues with shared dependencies
+
 ## 1.18.1
 
 ### Patch Changes